Create util method
